Electronic Mock Examination in an Introductory Computer Science Course
A mock examination is an exam that does not count for credit. It is taken before an official examination and gives students the opportunity to practice for the later, important exam. The mock examination provides students with information on their actual learning progress and gives them in this way the opportunity to fill knowledge gaps. An electronic mock examination is the online variant usually taken unsupervised in the students’ own time. In this work we present the setup of an electronic mock exam in the course “Basics of Computer Science”. We discuss question types other than multiple-choice that allow for fully automated marking. Furthermore, we present an in-depth analysis of the results of the mock exam. These results demonstrate that students consider the electronic mock
exam as valuable help for exam preparation. Half of the students take the mock exam even more than once and use different
strategies to make the most ou